pyproject.toml
[tool.poetry]
name = "Api Rest Django"
version = "0.1.0"
description = ""
authors = ["Nicola Gernone"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
model-bakery = "^1.7.0"
djangorestframework-xml = "^2.0.0"

[tool.poetry.dev-dependencies]
coverage = "^6.3.1"
pandas = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

